refactor(arch): lift runOneBackfillBatch to ~/lib/embed-backfill (CF-002)
The pipeline-consumer queue handler imported the embedding batch helper
directly from the admin route file, which inverted the dependency
direction (queue depending on the HTTP/auth layer). Move the helper
and countRemaining into ~/lib/embed-backfill; the route stays a thin
HTTP wrapper; the queue consumer imports the lib directly. The
pipeline-consumer test's vi.mock target updates accordingly.

Phase decomposition of pipeline-consumer.ts (the cosmetic part of
CF-002) is deferred — splitting the 7 phase handlers into per-file
modules without local builds is high-risk theater that the line-count
smell does not justify.
